rename config_poll module to polling
